OSN-747. Support Azure RI
