+++ risp/risp_worker/pyproject.toml
@@ -5,6 +5,7 @@ requires-python = ">=3.12.11,<3.13"
 dependencies = [
     "boto3==1.34.7",
     "clickhouse-connect==0.8.15",
+    "cloud-adapter",
     "config-client",
     "cryptography==44.0.1",
     "kombu==5.3.4",
@@ -12,11 +13,14 @@ dependencies = [
     "pymongo==4.6.3",
     "pyyaml==6.0.1",
     "restapi-client",
+    "insider-client",
 ]
 
 [tool.uv.sources]
 config-client = { path = "../../optscale_client/config_client" }
+insider-client = { path = "../../optscale_client/insider_client" }
 restapi-client = { path = "../../optscale_client/rest_api_client" }
+cloud-adapter = { path = "../../tools/cloud_adapter" }
 optscale-exceptions = { path = "../../tools/optscale_exceptions" }
 
 [dependency-groups]
