+++ tools/cloud_adapter/setup.py
@@ -29,7 +29,8 @@ requirements = [
     "azure-mgmt-consumption==8.0.0",
     "azure-mgmt-monitor==0.10.0",
     "azure-mgmt-commerce==1.0.1",
-    "msrest==0.6.6",
+    "azure-mgmt-reservations==2.3.0",
+    "msrest==0.7.1",
     "urllib3==2.5.0",
     "azure-identity==1.16.1",
     "oauthlib==3.2.2",
